+++ backends-velox/src/main/scala/io/glutenproject/execution/ShuffledHashJoinExecTransformer.scala
@@ -16,18 +16,13 @@
  */
 package io.glutenproject.execution
 
-import io.glutenproject.utils.Iterators
-
-import org.apache.spark.{broadcast, SparkContext}
+import io.substrait.proto.JoinRel
 import org.apache.spark.sql.catalyst.expressions._
 import org.apache.spark.sql.catalyst.optimizer.{BuildLeft, BuildRight, BuildSide}
 import org.apache.spark.sql.catalyst.plans._
-import org.apache.spark.sql.execution.{FilterExec, SparkPlan}
 import org.apache.spark.sql.execution.aggregate.BaseAggregateExec
 import org.apache.spark.sql.execution.joins.BuildSideRelation
-import org.apache.spark.sql.vectorized.ColumnarBatch
-
-import io.substrait.proto.JoinRel
+import org.apache.spark.sql.execution.{FilterExec, SparkPlan}
 
 case class ShuffledHashJoinExecTransformer(
     leftKeys: Seq[Expression],
@@ -155,20 +150,6 @@ case class ShuffledHashJoinExecTransformer(
       newLeft: SparkPlan,
       newRight: SparkPlan): ShuffledHashJoinExecTransformer =
     copy(left = newLeft, right = newRight)
-}
-
-case class VeloxBroadcastBuildSideRDD(
-    @transient private val sc: SparkContext,
-    broadcasted: broadcast.Broadcast[BuildSideRelation])
-  extends BroadcastBuildSideRDD(sc, broadcasted) {
-
-  override def genBroadcastBuildSideIterator(): Iterator[ColumnarBatch] = {
-    val relation = broadcasted.value.asReadOnlyCopy()
-    Iterators
-      .wrap(relation.deserialized)
-      .recyclePayload(batch => batch.close())
-      .create()
-  }
 }
 
 case class GlutenBroadcastHashJoinExecTransformer(
